fix(brain): PATH seeding, preflight warning severity, fallback call fidelity, client split (review)
- The CLI subprocess PATH is seeded with the detector's fallback dirs
  (single source: AgentCLIDetector.fallbackDirectories), so a shim
  found in one dir resolves its interpreter from another — detection
  and execution now see the same world.
- New UserFacingError severity .warning (alert, never stop): the CLI
  preflight refusals (brainCLIMissing / brainCLINotSignedIn) use it,
  so an in-place restart that trips the guard no longer tears down the
  live session it was supposed to leave untouched.
- The forced-speak fallback records the lines actually shown in its
  RawToolCall arguments instead of {}, so replayed history reports the
  turn faithfully and satisfies the speak schema.
- CLIBrainClient split into purpose-named files (core respond flow /
  +Conversation / +Invocation / +ReplyParsing), none past ~190 lines.

Live-verified on both CLIs after the changes.
