+++ Tests/JarvisCoreTests/Diagnostics/UserFacingErrorTests.swift
@@ -7,6 +7,12 @@ import Testing
         #expect(UserFacingError.Severity.fatal.stopsSession)
     }
 
+    @Test func warningAlertsWithoutStopping() {
+        // The preflight severity: the failed thing never started, so a live session must survive.
+        #expect(UserFacingError.Severity.warning.showsAlert)
+        #expect(!UserFacingError.Severity.warning.stopsSession)
+    }
+
     @Test func degradedNeitherAlertsNorStops() {
         #expect(!UserFacingError.Severity.degraded.showsAlert)
         #expect(!UserFacingError.Severity.degraded.stopsSession)
